Fix - Fix format exceptions for double values in Response

+++ app/src/main/java/com/jossprogramming/weatherapp/network/models/Astro.kt
@@ -1,7 +1,7 @@
 package com.jossprogramming.weatherapp.network.models
 
 data class Astro(
-    val moon_illumination: Int,
+    val moon_illumination: Double,
     val moon_phase: String,
     val moonrise: String,
     val moonset: String,
